+++ plugin-dev/Source/Sentry/Private/SentrySubsystem.cpp
@@ -289,12 +289,12 @@ void USentrySubsystem::LogFatalWithAttributes(const FString& Message, const TMap
 	AddLog(Message, ESentryLevel::Fatal, Attributes, Category);
 }
 
-void USentrySubsystem::AddCount(const FString& Key, float Value, FSentryMeasurementUnit Unit)
+void USentrySubsystem::AddCount(const FString& Key, int32 Value, FSentryMeasurementUnit Unit)
 {
 	AddCountWithAttributes(Key, Value, Unit, TMap<FString, FSentryVariant>());
 }
 
-void USentrySubsystem::AddCountWithAttributes(const FString& Key, float Value, FSentryMeasurementUnit Unit, const TMap<FString, FSentryVariant>& Attributes)
+void USentrySubsystem::AddCountWithAttributes(const FString& Key, int32 Value, FSentryMeasurementUnit Unit, const TMap<FString, FSentryVariant>& Attributes)
 {
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
 	{
